putting BoR rules in the right module.... whoops
